+++ TabletopTweaks/NewContent/MythicAbilities/ArmoredMight .cs
@@ -6,7 +6,7 @@ using TabletopTweaks.NewComponents.AbilitySpecific;
 using TabletopTweaks.Utilities;
 
 namespace TabletopTweaks.NewContent.MythicAbilities {
-    class ArmoredMight {
+    static class ArmoredMight {
         public static void AddArmoredMight() {
             var MythicAbilitySelection = Resources.GetBlueprint<BlueprintFeatureSelection>("ba0e5a900b775be4a99702f1ed08914d");
             var ExtraMythicAbilityMythicFeat = Resources.GetBlueprint<BlueprintFeatureSelection>("8a6a511c55e67d04db328cc49aaad2b8");