+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Dynamic item naming of armor no longer includes the enhancement bonus of the armor when it does not deviate from the original value
 * Starlight is no longer affected by true sight as it is not an illusion effect
 * Updated Horsemaster prerequisites
+* Added Abundant Bombs mythic ability
 
 ## Version 1.10.4
 * Now works with 1.1.6e